pyproject.toml
[tool.poetry]
name = "homework30-django-auth-permissions"
version = "0.1.0"
description = ""
authors = ["MichaelRodionov <106465054+MichaelRodionov@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "homework30_django_auth_permissions"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
pillow = "^9.4.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
